Removed jQuery dependency (#57)
* Removed jQuery dependency

With jQuery required so explicitly, this pretty much forces the user to have assets at the top of the file, else jQuery will be undefined. By removing the dependency on jQuery, this is still cross-browser compatible but now allows for JS assets to be output anywhere else in the page.

* Small bug

Small bugfix with references to values inside closures.

* Checking for button's presence

Avoid checking for the search button if the setting says not to show it.